+++ tests/unit/test_optimisation_problems/test_OCP_balance.py
@@ -22,7 +22,7 @@ class TestOCPBalance(unittest.TestCase):
                 ["data_fit"], ["data_ref1", "data_ref2"]
             )
 
-    def test_setup_cost_function(self):
+    def test_setup_objective_function(self):
         data_ref = [
             pd.DataFrame({0: [0.1, 0.3, 0.5, 0.7, 0.9], 1: [5, 4, 3, 2, 1]}),
             pd.DataFrame({0: [0.1, 0.3, 0.5, 0.7, 0.9], 1: [6, 5, 4, 3, 2]}),
@@ -35,14 +35,14 @@ class TestOCPBalance(unittest.TestCase):
         ]
         optimisation_problem = pbparam.OCPBalance(data_fit, data_ref)
 
-        optimisation_problem.setup_cost_function()
+        optimisation_problem.setup_objective_function()
 
         # Check bounds are correct
         self.assertEqual(optimisation_problem.x0, [-0.25, 0.25])
         self.assertEqual(optimisation_problem.bounds, [(-0.275, 1.1), (-0.1, 0.275)])
 
         # Check cost function is zero at theoretical optimal
-        self.assertAlmostEqual(optimisation_problem.cost_function([-0.1, 0.2]), 0)
+        self.assertAlmostEqual(optimisation_problem.objective_function([-0.1, 0.2]), 0)
 
         # Test increasing fit data
         data_fit = [
@@ -51,19 +51,19 @@ class TestOCPBalance(unittest.TestCase):
         ]
         optimisation_problem = pbparam.OCPBalance(data_fit, data_ref)
 
-        optimisation_problem.setup_cost_function()
+        optimisation_problem.setup_objective_function()
 
         # Check bounds are correct
         self.assertEqual(optimisation_problem.x0, [1.25, -0.25])
         self.assertEqual(optimisation_problem.bounds, [(-0.1, 1.375), (-0.275, 0.1)])
 
         # Check cost function is zero at theoretical optimal
-        self.assertAlmostEqual(optimisation_problem.cost_function([1.1, -0.2]), 0)
+        self.assertAlmostEqual(optimisation_problem.objective_function([1.1, -0.2]), 0)
 
         # Test data type error
         optimisation_problem = pbparam.OCPBalance(["data_fit"], ["data_ref"])
         with self.assertRaisesRegex(TypeError, "data_ref elements must"):
-            optimisation_problem.setup_cost_function()
+            optimisation_problem.setup_objective_function()
 
     def test_plot(self):
         data_ref = [
@@ -76,7 +76,7 @@ class TestOCPBalance(unittest.TestCase):
         ]
         optimisation_problem = pbparam.OCPBalance(data_fit, data_ref)
 
-        optimisation_problem.setup_cost_function()
+        optimisation_problem.setup_objective_function()
 
         fig = optimisation_problem._plot([-0.1, 0.2])
         ax = fig.axes[0]