+++ tests/unit/test_optimisers/test_scipy_differential_evolution.py
@@ -20,7 +20,7 @@ class TestScipyDifferentialEvolution(unittest.TestCase):
 
     def test_optimiser(self):
         opt = pbparam.BaseOptimisationProblem()
-        opt.cost_function = parabola
+        opt.objective_function = parabola
         opt.x0 = 2
         opt.bounds = [[-10, 10]]
 
@@ -30,7 +30,7 @@ class TestScipyDifferentialEvolution(unittest.TestCase):
 
     def test_optimiser_multiple_workers(self):
         opt = pbparam.BaseOptimisationProblem()
-        opt.cost_function = parabola
+        opt.objective_function = parabola
         opt.x0 = 2
         opt.bounds = [[-10, 10]]